requirements.txt
gym==0.9.3
statsd==3.2.1
requests
python-dotenv==0.8.2
tensorflow==1.5.1